% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{apply_grouped}
\alias{apply_grouped}
\title{Functional for precomputed D4 Orbit Groups}
\usage{
apply_grouped(spec_mat, groups, fun, ...)
}
\arguments{
\item{spec_mat}{N x M matrix}

\item{groups}{List of indices (orbits)}

\item{fun}{A function that takes a vector and returns a vector of the same length}

\item{...}{Additional arguments passed to fun}
}
\description{
Functional for precomputed D4 Orbit Groups
}
