% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/I_smooth.R
\name{evaluate.grid_function}
\alias{evaluate.grid_function}
\title{evaluate a grid function object}
\usage{
\method{evaluate}{grid_function}(f, N = 30, M = 30)
}
\arguments{
\item{f}{the grid_function object}

\item{N}{integer for number of rows to generate}

\item{M}{integer for number of columns to generate}
}
\description{
Evaluates a function on \eqn{[-\pi, \pi]^2}
}
