% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/I_smooth.R
\name{evaluate}
\alias{evaluate}
\title{evaluate generic}
\usage{
evaluate(f, N, M)
}
\arguments{
\item{f}{function}

\item{N}{integer for rows of output matrix}

\item{M}{integer for cols of output matrix}
}
\description{
Evaluates functions on a grid
}
\examples{
f <- function(x, y) x + y
class(f) <- "grid_function"
vals <- evaluate(f, 4, 4)
dim(vals$z)
}
