% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridMA.R
\name{gridMA}
\alias{gridMA}
\title{Simulate a spatial MA process}
\usage{
gridMA(N, M, K, distribution = "normal")
}
\arguments{
\item{N}{rows}

\item{M}{columns of grid}

\item{K}{numeric matrix of kernel weights}

\item{distribution}{Type of distribution must be one of "normal", "uniform", "cauchy", "chisq"}
}
\value{
matrix with N rows and M columns
}
\description{
To validate the implementation we need data generated under both \eqn{H_0}
and the alternative. We therefore simulate a simple spatial \eqn{MA(q)} process.
The idea is to interpret lag coefficients as distance based influence on
\eqn{X_{ij}}, which can be implemented via a 2D convolution. The resulting
process is stationary with an isotropic covariance function.
}
\examples{
set.seed(1)
K <- MA_coef_all(0.3)
x <- gridMA(25, 25, K)
image(x)
}
