% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/I.R
\name{I}
\alias{I}
\title{2D Periodogramm}
\usage{
I(x)
}
\arguments{
\item{x}{matrix with data from stationary spatial process}
}
\value{
martix with results
}
\description{
Computes 2D Periodogramm. Consider the discrete Fourier transform of the sample:
\deqn{J({\omega}) = \frac{1}{\sqrt{2\pi NM}} \sum_{s\in S} x(s)\,\exp\!\left(-i\langle s,{\omega}\rangle\right)}

Then the 2D periodogram is given by:
\deqn{I(\omega_{kl}) = J(\omega_{kl})\,\overline{J(\omega_{kl})} = \big|J(\omega_{kl})\big|^2}
}
