% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/I_smooth.R
\name{I_smooth}
\alias{I_smooth}
\title{Smoothed Periodogram}
\usage{
I_smooth(I, hr = 0.2, hc = 0.2)
}
\arguments{
\item{I}{periodogram results}

\item{hr}{rowwise bandwidth}

\item{hc}{columnwise bandwidth}

\item{kernel}{kernel used to smooth}
}
\description{
Computes smoothed Periodogram
}
