% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JPtest.R
\name{JPtest}
\alias{JPtest}
\title{Jentsch Pauly Test}
\usage{
JPtest(x, y, B, alpha, print_result = TRUE, hypothesis = "equality")
}
\arguments{
\item{x}{first sample. Can be numeric vector or matrix}

\item{y}{second sample. Can be numeric vector or matrix (but must in it's type coincide with x). If isotropy is tested, this is disregarded.}

\item{B}{Number of iterations for resampling (The more the better). Can be numeric but must be a whole number}

\item{alpha}{Level of significance. Must be numeric value in (0,1]}

\item{print_result}{Boolean whether to print the result or just return an object cotaining the result}

\item{hypothesis}{Can be one of `"equality"` or `"isotropy"`}
}
\description{
Compute resampling test for equality of spectral densities or other qualities of the spectral densities.
Can be applied to both vectors and matrices. For more information see DOI: 10.3150/13-BEJ584.
}
\examples{
# simulating grid-data under H0
K0 <- MA_coef_all(.3)
x <- gridMA(25, 25, K0)
y <- gridMA(25, 25, K0)

# applying the test with 100 iterations and significance 5\%
JPtest(x, y, 100, .05)

}
