% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{k_2d_bp}
\alias{k_2d_bp}
\title{Construct Kernel from Barlett-Priestley window}
\usage{
k_2d_bp(N, M, hr = 0.5, hc = 0.5)
}
\description{
Used in the fast version of phi_n_star to construct a kernel for the convolution
}
