% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{MA_coef_all}
\alias{MA_coef_all}
\title{Generate Coefficient Matrix for MA Process}
\usage{
MA_coef_all(value, size = 3)
}
\arguments{
\item{value}{value to be added surounding}

\item{size}{size of the matrix}
}
\description{
Generate Coefficient Matrix for MA Process
}
\examples{
MA_coef_all(0.5)
}
