% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{MA_coef_col}
\alias{MA_coef_col}
\title{Generate Coefficient Matrix for MA Process}
\usage{
MA_coef_col(value, size = 3)
}
\arguments{
\item{value}{value to be added in column}

\item{size}{size of the matrix}
}
\description{
Generate Coefficient Matrix for MA Process
}
