% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt3.R
\name{new_periodoTestResult}
\alias{new_periodoTestResult}
\title{Constructor for periodoTestResult object}
\usage{
new_periodoTestResult(test_value, p_value, decision, alpha, hypothesis)
}
\description{
Constructor for periodoTestResult object
}
