% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testResult.R
\name{new_testResult}
\alias{new_testResult}
\title{Constructor for testResult object}
\usage{
new_testResult(Tn, Tn_star, decision, p_value, hypothesis, B, alpha, h1, h2)
}
\description{
Constructor for testResult object
}
