% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/I_smooth.R
\name{plot.I_smooth}
\alias{plot.I_smooth}
\title{Plot an smoothed periodogramm}
\usage{
\method{plot}{I_smooth}(x, ...)
}
\arguments{
\item{x}{I_smooth function}

\item{...}{additional arguments might become relevant in later versions}
}
\description{
This function plots a smoothed periodogramm.
}
