% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_true_MA.R
\name{plot.spectral_density}
\alias{plot.spectral_density}
\title{Plot a spectral density}
\usage{
\method{plot}{spectral_density}(x, ...)
}
\arguments{
\item{x}{spectral_density object}

\item{...}{additional arguments passed to the plot}
}
\description{
Plot a spectral density
}
\examples{
f <- function(x, y) cos(x) + cos(y)
class(f) <- c("spectral_density", "grid_function")
if (interactive()) {
  plot(f)
}
}
