% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testResult.R
\name{print.testResult}
\alias{print.testResult}
\title{Print generic for testResult S3 class}
\usage{
\method{print}{testResult}(x, ...)
}
\arguments{
\item{x}{testResult object}

\item{...}{additional parameters for printing (might become relevant in future versions)}
}
\description{
Print generic for testResult S3 class
}
