% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testResult.R
\name{summary.testResult}
\alias{summary.testResult}
\title{summary generic for S3 testResult class}
\usage{
\method{summary}{testResult}(object, ...)
}
\arguments{
\item{object}{testResult object}

\item{...}{additional parameters for printing (might become relevant in future versions)}
}
\description{
summary generic for S3 testResult class
}
