% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.periodo.R
\name{test.periodo}
\alias{test.periodo}
\title{Periodogram Test for Assessing Equality of Spectral Densities}
\usage{
test.periodo(x, y, alpha)
}
\arguments{
\item{x}{numeric matrix of dims N, M with no NA values (representing the first lattice data sample)}

\item{y}{numeric matrix of dims N, M with no NA values (representing the second lattice data sample)}

\item{alpha}{numeric value in \eqn{(0,1]}}
}
\value{
An object of type periodoTestResult
}
\description{
\eqn{\varphi_n^p} is an adaption of the test \eqn{T_3} from the work by Scaccia and Martin (2005). The test compares individual deviations of the periodogram ordinates
 \eqn{I_x} and \eqn{I_y} from the total amplitude of \eqn{I_x + I_y}. We define \eqn{G(\omega_{kl})} as the comparison value:
 \deqn{G(\omega_{kl}) = \frac{I_x(\omega_{kl})-I_y(\omega_{kl})}{I_x(\omega_{kl}) + I_y(\omega_{kl})}}
 The test statistic \eqn{PT_3} is then given by:
 \deqn{PT_3 = \sqrt{12n}(\overline{|G|-1/2}) \overset d\longrightarrow \mathcal N(0,1)}
 Critical values are drawn from the standard normal distirbution.
}
