% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridMA.R
\name{gridMA}
\alias{gridMA}
\title{simulating spatial MA process}
\usage{
gridMA(N, M, K, padding = nrow(K)\%/\%2)
}
\arguments{
\item{N}{rows}

\item{M}{columns of grid}
}
\value{
matrix with N rows and M columns
}
\description{
Um unsere Implementierung validieren zu können, brauchen wir Daten, welche sowohl unter \eqn{H_0} als auch unter der Alternative erzeugt wurden. Wie bereits in der Einleitung erwähnt, wollen wir alles so simpel wie möglich halten. Der Prozess ist daher eine einfache Linearkombination von normalverteilten Variablen und stark von der Simulation eines \eqn{MA(q)} Prozesses inspiriert. Der Grundgedanke hierbei ist, dass wir die Koeffizienten \eqn{\varphi_k} des Lag-Polynoms reinterpretieren von "linearer Einfluss von Residuen zu Lag \eqn{k}" zu "linearer Einfluss von Residuen zum Abstand \eqn{k}". Also heißt dass für \eqn{X_{ij}}, dass jedes Residuum mit Abstand \eqn{k} einen linearen Einfluss von \eqn{\varphi_k} auf \eqn{X_{ij}} nimmt. Dieser Prozess lässt sich mit einer 2D Konvolution umsetzen und ist nicht nur stationär, sondern hat auch eine isotrope Kovarianzfunktion.
}
