% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/I_smooth.R
\name{I_smooth}
\alias{I_smooth}
\title{Smoothed Periodogram}
\usage{
I_smooth(I, hr = 0.2, hc = 0.2)
}
\arguments{
\item{I}{periodogram results}

\item{hr}{rowwise bandwidth}

\item{hc}{columnwise bandwidth}
}
\description{
Computes smoothed Periodogram
}
\examples{
set.seed(1)
x <- matrix(rnorm(9), nrow = 3)
smoothed_periodo <- I_smooth(I(x))
plot(smoothed_periodo)

}
