% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi_n_star.R
\name{phi_n_star_mv}
\alias{phi_n_star_mv}
\title{\eqn{\varphi_n^*} for \eqn{p=1} and arbitray \eqn{q}.}
\usage{
phi_n_star_mv(X, B, alpha, h1 = 0.14, h2 = 0.14)
}
\description{
\eqn{\varphi_n^*} for \eqn{p=1} and arbitray \eqn{q}.
}
