% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/I_smooth.R
\name{plot.I_smooth}
\alias{plot.I_smooth}
\title{Plot an smoothed periodogramm}
\usage{
\method{plot}{I_smooth}(I_s, ...)
}
\arguments{
\item{I_s}{I_smooth function}
}
\description{
This function plots a smoothed periodogramm.
}
