% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt3.R
\name{print.periodoTestResult}
\alias{print.periodoTestResult}
\title{summary generic for S3 testResult class}
\usage{
\method{print}{periodoTestResult}(x, ...)
}
\description{
summary generic for S3 testResult class
}
