% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testResult.R
\name{print.testResult}
\alias{print.testResult}
\title{Print generic for testResult S3 class}
\usage{
\method{print}{testResult}(x, ...)
}
\arguments{
\item{x}{testResult object}

\item{...}{additional parameters for printing (might become relevant in future versions)}
}
\description{
Print generic for testResult S3 class
}
\examples{
res <- new_testResult(1, c(0.5, 1.5), 0, 0.5, "equality", 2, 0.05, 0.1, 0.1)
print(res)
}
