% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt3.R
\name{summary.periodoTestResult}
\alias{summary.periodoTestResult}
\title{summary generic for S3 testResult class}
\usage{
\method{summary}{periodoTestResult}(object, ...)
}
\arguments{
\item{object}{periodoTestResult Object}

\item{...}{additional parameters (might be used in later versions)}
}
\description{
summary generic for S3 testResult class
}
