% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testResult.R
\name{summary.testResult}
\alias{summary.testResult}
\title{summary generic for S3 testResult class}
\usage{
\method{summary}{testResult}(object, ...)
}
\arguments{
\item{object}{testResult object}

\item{...}{additional parameters for printing (might become relevant in future versions)}
}
\description{
summary generic for S3 testResult class
}
\examples{
res <- new_testResult(1, c(0.5, 1.5), 0, 0.5, "equality", 2, 0.05, 0.1, 0.1)
summary(res)
}
